🔨 피드백 반영 : 파사드 메서드 변환 로직 수정 및 서비스 리팩토링(가독성 향상)
