+++ src/main/java/kr/hhplus/be/server/presentation/concert/dto/ConcertSeatResponse.java
@@ -1,5 +1,6 @@
 package kr.hhplus.be.server.presentation.concert.dto;
 
+import kr.hhplus.be.server.domain.concert.ConcertSeat;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -10,4 +11,14 @@ public class ConcertSeatResponse {
     private Long seatNo;
     private String status;
     private Long price;
+
+    public static ConcertSeatResponse fromEntity(ConcertSeat concertSeat) {
+        return ConcertSeatResponse
+                .builder()
+                .id(concertSeat.getId())
+                .price(concertSeat.getPrice())
+                .seatNo(concertSeat.getSeatNo())
+                .status(concertSeat.getStatus())
+                .build();
+    }
 }
