+++ src/test/java/kr/hhplus/be/server/unit/ReservationFacadeTest.java
@@ -8,7 +8,7 @@ import kr.hhplus.be.server.domain.reservation.ReservationService;
 import kr.hhplus.be.server.domain.user.User;
 import kr.hhplus.be.server.domain.user.UserService;
 import kr.hhplus.be.server.application.reservation.ReservationFacade;
-import kr.hhplus.be.server.presentation.reservation.dto.ReservationResponse;
+import kr.hhplus.be.server.application.reservation.dto.ReservationResponse;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InOrder;
@@ -98,7 +98,7 @@ public class ReservationFacadeTest {
         inOrder.verify(userService).makePayment(any(), any());
 
         // 4. 좌석 상태 변경
-        inOrder.verify(concertService).saveConcertSeat(any());
+        inOrder.verify(concertService).changeConcertSeatCompleted(any());
 
         // 5. 예약 상태 변경
         inOrder.verify(reservationService).completeReservation(any());
